Start cleaning up cruft and TODOs and move inner classes to top level.
