+++ Src/java/cql-to-elm/src/main/java/org/cqframework/cql/cql2elm/LibraryManager.java
@@ -11,7 +11,6 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.StringJoiner;
 import org.apache.commons.lang3.StringUtils;
 import org.cqframework.cql.cql2elm.model.CompiledLibrary;
 import org.cqframework.cql.elm.serializing.ElmLibraryReaderFactory;
@@ -168,6 +167,7 @@ public class LibraryManager {
         return this.resolveLibrary(libraryIdentifier, errors, CacheMode.READ_WRITE);
     }
 
+    // LUKETODO:  call the multilib method here
     public CompiledLibrary resolveLibrary(
             VersionedIdentifier libraryIdentifier, List<CqlCompilerException> errors, CacheMode cacheMode) {
         if (libraryIdentifier == null) {
@@ -189,7 +189,6 @@ public class LibraryManager {
         var compileLibraryResult = compileLibrary(libraryIdentifier);
         library = compileLibraryResult.compiledLibrary();
         if (!hasErrors(compileLibraryResult.errors()) && cacheMode == CacheMode.READ_WRITE) {
-            logger.info("1234: adding library to cache: {}", libraryIdentifier.getId());
             compiledLibraries.put(libraryIdentifier, library);
         } else {
             // LUKETODO: can we just return the result record instead ?
@@ -206,11 +205,11 @@ public class LibraryManager {
         return resolveLibraries(libraryIdentifiers, errorsById, CacheMode.READ_WRITE);
     }
 
-    // LUKETODO: rethink this to allow partial errors and let the other libraries through?
     public List<CompiledLibrary> resolveLibraries(
             List<VersionedIdentifier> libraryIdentifiers,
             Map<VersionedIdentifier, List<CqlCompilerException>> errorsById,
             CacheMode cacheMode) {
+
         if (libraryIdentifiers == null || libraryIdentifiers.isEmpty()) {
             throw new IllegalArgumentException("libraryIdentifier is null or empty.");
         }
@@ -221,14 +220,9 @@ public class LibraryManager {
             throw new IllegalArgumentException("at least one libraryIdentifier Id is null");
         }
 
-        logger.info("1234: resolveLibraries called with {} libraries", libraryIdentifiers.size());
-
         // LUKETODO:  do we need to order these?
         var libs = new ArrayList<CompiledLibrary>();
 
-        // LUKETODO:  this is WRONG:  if we find only 1-3 libraries in the cache, we should not return the whole thing,
-        // we should compile the others
-        // LUKETODO:  we don't pull from the cache because the versions don't match
         if (cacheMode != CacheMode.NONE) {
 
             // Ensure that cache retrieved libraries are in the same order as the input identifiers so we
@@ -238,15 +232,6 @@ public class LibraryManager {
                     .map(compiledLibraries::get)
                     .toList();
 
-            logger.info(
-                    "1234: libraries found in cache: {}",
-                    libraries.stream()
-                            .map(CompiledLibrary::getIdentifier)
-                            .map(VersionedIdentifier::getId)
-                            .toList());
-
-            //            if (!libraries.isEmpty()) {
-            // LUKETODO:  simple check:  should we match the identifiers as well?
             if (libraries.size() == libraryIdentifiers.size()) {
                 return libraries;
             }
@@ -255,23 +240,19 @@ public class LibraryManager {
         }
 
         for (VersionedIdentifier libraryIdentifier : libraryIdentifiers) {
-            // LUKETODO: check if the library is already in the cache, if so, skip it
             if (libs.stream().map(CompiledLibrary::getIdentifier).toList().contains(libraryIdentifier)) {
-                logger.info("1234: library {} already in cache, skipping compilation", libraryIdentifier.getId());
+                logger.debug("library {} already in cache, skipping compilation", libraryIdentifier.getId());
                 continue;
             }
 
             var compiledlibraryResult = compileLibrary(libraryIdentifier);
 
-            logger.info("1234: compile library result: {}", compiledlibraryResult);
-
-            // LUKETODO:  we need to add to libs irrespective of CacheMode.READ_WRITE:  is there a way to test
             // If we have any errors, ignore the compiled library altogether just like in the single lib case
             if (!hasErrors(compiledlibraryResult.errors())) {
                 // add to returned libs regardless of cache mode
                 libs.add(compiledlibraryResult.compiledLibrary());
                 if (cacheMode == CacheMode.READ_WRITE) {
-                    logger.info("1234: adding library to cache: {}", libraryIdentifier.getId());
+                    logger.debug("adding library to cache: {}", libraryIdentifier.getId());
                     compiledLibraries.put(libraryIdentifier, compiledlibraryResult.compiledLibrary());
                 }
             }
@@ -283,17 +264,6 @@ public class LibraryManager {
         }
 
         return List.copyOf(libs);
-    }
-
-    // LUKETODO: put this in a separate file
-    record CompiledLibraryResult(CompiledLibrary compiledLibrary, List<CqlCompilerException> errors) {
-        @Override
-        public String toString() {
-            return new StringJoiner(", ", CompiledLibraryResult.class.getSimpleName() + "[", "]")
-                    .add("compiledLibrary=" + compiledLibrary.getIdentifier())
-                    .add("errors=" + errors)
-                    .toString();
-        }
     }
 
     private CompiledLibraryResult compileLibrary(VersionedIdentifier libraryIdentifier) {
@@ -365,70 +335,6 @@ public class LibraryManager {
     }
 
     // LUKETODO:  look at newly unused private methods and get rid of them
-    private CompiledLibrary compileLibrary(VersionedIdentifier libraryIdentifier, List<CqlCompilerException> errors) {
-
-        CompiledLibrary result = null;
-        String libraryPath = NamespaceManager.getPath(libraryIdentifier.getSystem(), libraryIdentifier.getId());
-
-        if (!this.cqlCompilerOptions.getEnableCqlOnly()) {
-            result = tryCompiledLibraryElm(libraryIdentifier, this.cqlCompilerOptions);
-            if (result != null) {
-                validateIdentifiers(libraryIdentifier, result, libraryPath);
-
-                sortStatements(result);
-                return result;
-            }
-        }
-
-        try {
-            InputStream cqlSource = librarySourceLoader.getLibrarySource(libraryIdentifier);
-            if (cqlSource == null) {
-                throw new CqlIncludeException(
-                        String.format(
-                                "Could not load source for library %s, version %s, namespace uri %s.",
-                                libraryIdentifier.getId(),
-                                libraryIdentifier.getVersion(),
-                                libraryIdentifier.getSystem()),
-                        libraryIdentifier.getSystem(),
-                        libraryIdentifier.getId(),
-                        libraryIdentifier.getVersion());
-            }
-
-            CqlCompiler compiler = new CqlCompiler(
-                    namespaceManager.getNamespaceInfoFromUri(libraryIdentifier.getSystem()), libraryIdentifier, this);
-            compiler.run(cqlSource);
-            if (errors != null) {
-                errors.addAll(compiler.getExceptions());
-            }
-
-            result = compiler.getCompiledLibrary();
-
-            if (result == null) {
-                throw new CqlIncludeException(
-                        String.format(
-                                "Could not load source for library %s, version %s.",
-                                libraryPath, libraryIdentifier.getVersion()),
-                        libraryIdentifier.getSystem(),
-                        libraryIdentifier.getId(),
-                        libraryIdentifier.getVersion());
-            }
-
-            validateIdentifiers(libraryIdentifier, result, libraryPath);
-
-        } catch (IOException e) {
-            throw new CqlIncludeException(
-                    String.format(
-                            "Errors occurred translating library %s, version %s.",
-                            libraryPath, libraryIdentifier.getVersion()),
-                    libraryIdentifier.getSystem(),
-                    libraryIdentifier.getId(),
-                    libraryIdentifier.getVersion(),
-                    e);
-        }
-
-        sortStatements(result);
-        return result;
-    }
 
     private void validateIdentifiers(
             VersionedIdentifier libraryIdentifier, CompiledLibrary result, String libraryPath) {
